+++ .gitignore
@@ -1,10 +1,3 @@
+.vercel
+.env
 node_modules
-dist
-.wrangler
-.dev.vars
-
-# Change them to your taste:
-package-lock.json
-yarn.lock
-pnpm-lock.yaml
-bun.lockb
