+++ drizzle.config.ts
@@ -3,7 +3,7 @@ import * as dotenv from "dotenv";
 dotenv.config();
 
 export default {
-  schema: "./src/db/schema.ts",
+  schema: "./db/schema.ts",
   out: "./drizzle/migrations",
   driver: "turso",
   dbCredentials: {
